Added dev configurations for ues effluent and recycling
This applies to both angular and nest apps
