+++ apps/ues-effluent-angular/src/environments/definitions.ts
@@ -1,3 +1,12 @@
+import { SearchSource, SearchSourceQueryParamsProperties } from '@tamu-gisc/search';
+import { LayerSource, LegendItem } from '@tamu-gisc/common/types';
+
+import { tiers } from './tier-dictionary';
+
+import { Popups } from '@tamu-gisc/aggiemap';
+
+import esri = __esri;
+
 export const Connections = {
   basemapUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/TAMU_BaseMap/MapServer',
   inforUrl: 'https://gis.tamu.edu/arcgis/rest/services/FCOR/MapInfo_20190529/MapServer',
@@ -16,5 +25,465 @@ export const Definitions = {
     layerId: 'buildings-layer',
     name: 'Buildings',
     url: `${Connections.basemapUrl}/1`
+  }
+};
+
+export const effluentZonesUrl = Connections.effluentZonesUrl;
+export const effluentSampleLocationsUrl = Connections.effluentSampleLocationsUrl;
+
+export const effluentTiers = tiers;
+
+const commonLayerProps = {
+  outFields: ['*'],
+  minScale: 100000,
+  maxScale: 0,
+  elevationInfo: { mode: 'relative-to-ground', offset: 1 } as esri.FeatureLayerElevationInfo,
+  popupEnabled: false
+};
+
+// Persistent layer definitions that will be processed by a factory and added to the map.
+export const LayerSources: LayerSource[] = [
+  {
+    type: 'feature',
+    id: Definitions.BUILDINGS.layerId,
+    title: Definitions.BUILDINGS.name,
+    url: Definitions.BUILDINGS.url,
+    popupComponent: Popups.BuildingPopupComponent,
+    listMode: 'hide',
+    loadOnInit: true,
+    visible: true,
+    layerIndex: 2,
+    native: {
+      ...commonLayerProps,
+      renderer: {
+        type: 'simple',
+        symbol: {
+          type: 'simple-fill',
+          style: 'solid',
+          color: [0, 0, 0, 0.01],
+          outline: {
+            width: '0'
+          }
+        }
+      }
+    }
+  },
+  {
+    type: 'feature',
+    id: 'sampling-zone-1',
+    title: 'Sampling Zone 1',
+    url: Connections.effluentZonesUrl,
+    popupComponent: Popups.BasePopupComponent,
+    listMode: 'show',
+    loadOnInit: true,
+    visible: false,
+    layerIndex: 3,
+    native: {
+      ...commonLayerProps,
+      definitionExpression: "Tier = '1'",
+      renderer: {
+        type: 'simple',
+        symbol: {
+          type: 'simple-fill',
+          color: [158, 85, 156, 0.4],
+          style: 'solid'
+        }
+      }
+    }
+  },
+  {
+    type: 'feature',
+    id: 'sampling-zone-2',
+    title: 'Sampling Zone 2',
+    url: Connections.effluentZonesUrl,
+    popupComponent: Popups.BasePopupComponent,
+    listMode: 'show',
+    loadOnInit: true,
+    visible: false,
+    layerIndex: 4,
+    native: {
+      ...commonLayerProps,
+      definitionExpression: "Tier = '2'",
+      renderer: {
+        type: 'simple',
+        symbol: {
+          type: 'simple-fill',
+          color: [167, 198, 54, 0.4],
+          style: 'solid'
+        }
+      }
+    }
+  },
+  {
+    type: 'feature',
+    id: 'sampling-zone-3',
+    title: 'Sampling Zone 3',
+    url: Connections.effluentZonesUrl,
+    popupComponent: Popups.BasePopupComponent,
+    listMode: 'show',
+    loadOnInit: true,
+    visible: true,
+    layerIndex: 5,
+    native: {
+      ...commonLayerProps,
+      definitionExpression: "Tier = '3'",
+      renderer: {
+        type: 'simple',
+        symbol: {
+          type: 'simple-fill',
+          color: [20, 158, 206, 0.4],
+          style: 'solid'
+        }
+      },
+      labelingInfo: [
+        {
+          labelExpressionInfo: {
+            expression: `$feature.SampleNumber`
+          },
+          symbol: {
+            type: 'text',
+            color: 'black',
+            haloSize: '1.5pt',
+            haloColor: 'white',
+            font: {
+              size: '8.5pt',
+              family: 'Ubuntu',
+              style: 'normal',
+              weight: 'bold'
+            }
+          }
+        }
+      ]
+    }
+  },
+  {
+    type: 'feature',
+    id: 'sampling-zone-4',
+    title: 'Sampling Zone 4',
+    url: Connections.effluentZonesUrl,
+    popupComponent: Popups.BasePopupComponent,
+    listMode: 'show',
+    loadOnInit: true,
+    visible: false,
+    layerIndex: 6,
+    native: {
+      ...commonLayerProps,
+      definitionExpression: "Tier = '4'",
+      renderer: {
+        type: 'simple',
+        symbol: {
+          type: 'simple-fill',
+          color: [237, 81, 81, 0.4],
+          style: 'solid'
+        }
+      }
+    }
+  },
+  {
+    type: 'feature',
+    id: 'sample-testing-locations',
+    title: 'Sample Testing Locations',
+    url: Connections.effluentSampleLocationsUrl,
+    popupComponent: Popups.BasePopupComponent,
+    listMode: 'show',
+    loadOnInit: true,
+    visible: true,
+    layerIndex: 7,
+    native: {
+      ...commonLayerProps,
+      definitionExpression: "Tier = '3'"
+    }
+  },
+  {
+    type: 'feature',
+    id: 'construction-layer',
+    title: 'Construction Zone',
+    url: `${Connections.constructionUrl}`,
+    category: 'Infrastructure',
+    listMode: 'show',
+    loadOnInit: false,
+    visible: true,
+    legendItems: [
+      {
+        id: 'construction-legend',
+        title: 'Construction Area',
+        src:
+          'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAENJREFUOI1jYaAyYKGZgf9DGf5TahjjagZGFnQBcg2DOYgFmyAlgLYupAYYYgaORsogNHA0Uig3kJIwhAUXC7oApQAAQ8kZ9+L+/N4AAAAASUVORK5CYII='
+      }
+    ],
+    popupComponent: Popups.ConstructionPopupComponent,
+    layerIndex: 1,
+    native: {
+      ...commonLayerProps
+    }
+  },
+  {
+    type: 'graphics',
+    id: 'selection-layer',
+    title: 'Selected Buildings',
+    category: 'Infrastructure',
+    listMode: 'hide',
+    loadOnInit: false,
+    visible: true,
+    popupComponent: Popups.BasePopupComponent
+  }
+];
+
+// Static legend items
+export const LegendSources: LegendItem[] = [
+  {
+    id: 'university-building-legend',
+    title: 'University Building',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAJBJREFUOI1jYaAyYIExAtnZGygxaP3Pnw1wAwPZ2RuE7azruYVFyDLsw/2HDCkvXhirKCvPgruQW1iEQUJJmWwXMv386cPAzOzDQlgpaWDUwFEDRw0cZgZ+uP+QbEO+vn3DwIts4PqfPxsCL1xg+HThAgMDAwODsISEMamGimtooLoQVuIyMDAwMDwkw7VQPQDNmyRazmV6EgAAAABJRU5ErkJggg=='
+  },
+  {
+    id: 'non-university-building-legend',
+    title: 'Non-University Building',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHFJREFUOI1jYaAyYIExmpqaGigxqK6urgFuYFNTU4OCBGe9EC8bWYY9ff2doampiaGurq4B7kIhXjYGZWk+sl348sMvBrgLqQlGDRw1cNTAYWbg09ffyTbk3edfqAbW1dU1NDU1wYsgcgBKAYssQCkAABa3ILBI1xApAAAAAElFTkSuQmCC'
+  },
+  {
+    id: 'off-campus-building-legend',
+    title: 'Off Campus Building',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGdJREFUOI3t1KENADEMA8CADhBc2H0802cnLxVUdYRHH7U0LXrVKAo4GbnI4ZTvIGk7EAALkKTVWh9VTWHuLiQFgEVDVZXWWrph712i4clc8IIX/Bno7mlkjLGCAIxkTFAmy8DOj9281XsgH/GbKEkAAAAASUVORK5CYII='
+  },
+  {
+    id: 'garage-parking-legend',
+    title: 'Garage Parking',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGtJREFUOI3t1LEOgCAMBNAOTMBgogNfezb3v4YQxdnJRldkMtzUdHi56Zx0jrsPkvoFAqAGktRlntYYfBO25SIkBYBawxi8pLQ0N9z3KtawZwY4wAH+DNxyaUaOer5BAErSJqglr4F9Pr7mAt2qILaQAOa8AAAAAElFTkSuQmCC'
+  },
+  {
+    id: 'restricted-legend',
+    title: 'Restricted 6am to 6pm',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAABHNCSVQICAgIfAhkiAAAAFtJREFUOI1jYKAyYIQxHB0dGygxaP/+/Q0MDAwMLDDDQkMC6+OfviDLsFmiwnBDmWCC5BrGwMDAkPb6LZzNhEcdWWDUwFEDRw0cZgbCiiBywEJpCTib6gUs1QEATW0WjAt6JIYAAAAASUVORK5CYII='
+  },
+  {
+    id: 'paved-surface-legend',
+    title: 'Paved Surface',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAGVJREFUOI3t1CEOACEMBMAKdF+A4aN7zf62AsMLTl0DtqAurGoqJqu2yOGU7yBpOxAAC5Ck1VofVU1h7i4kBYBFQ1WV1lq6Ye9douHJXPCCF/wZ6O5pZIyxggCMZExQJsvAzo/dvKA4IK9bweSrAAAAAElFTkSuQmCC'
+  },
+  {
+    id: 'sidewalk-legend',
+    title: 'Sidewalk',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAEtJREFUOI1jYaAyYIExPr641UCJQfwSag1wA6GG1VNi4McXtxj4JdQaWAgrJQ2MGjhq4KiBowbS0UB+CbWGjy9uUWQQSgGLLEApAAAvoxCdulpoFwAAAABJRU5ErkJggg=='
+  },
+  {
+    id: 'railroad-legend',
+    title: 'Railroad',
+    src:
+      'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABQAAAAUCAYAAACNiR0NAAAAAXNSR0IB2cksfwAAAAlwSFlzAAAOxAAADsQBlSsOGwAAAHhJREFUOI3tkrENgDAMBL+I9EUmYA7WYCfCTLBG5qBz/x0VRUAkAtIg8q3t+7dlh8pyDfgnIMnwBiQpJEBJI8mpMDQAWEnGYkIA0Xs/m9llM8kkTQlYRQnQzEJuHUk9gG5PeqiFBEhyKbnnzE4Jc3e5ow89dgM+1gZwZSYCkyusZQAAAABJRU5ErkJggg=='
+  },
+  {
+    id: 'effluent-green',
+    title: 'Positive',
+    src: './assets/images/legend/effluent-green.svg'
+  },
+  {
+    id: 'effluent-amber',
+    title: 'Inconclusive',
+    src: './assets/images/legend/effluent-amber.svg'
+  },
+  {
+    id: 'effluent-red',
+    title: 'Negative',
+    src: './assets/images/legend/effluent-reDefinitions.svg'
+  }
+];
+
+const commonQueryParams: Partial<SearchSourceQueryParamsProperties> = {
+  f: 'json',
+  resultRecordCount: 5,
+  outFields: '*',
+  outSR: 4326,
+  returnGeometry: true,
+  spatialRel: 'esriSpatialRelIntersects'
+};
+
+// Search sources used for querying features.
+export const SearchSources: SearchSource[] = [
+  {
+    source: 'building',
+    name: 'Building',
+    url: `${Connections.basemapUrl}/1`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['Number', 'BldgAbbr', 'BldgName'],
+        operators: ['LIKE', 'LIKE', 'LIKE'],
+        wildcards: ['includes', 'includes', 'includes'],
+        transformations: ['UPPER', 'UPPER', 'UPPER']
+      },
+      scoringWhere: {
+        keys: ['BldgName'],
+        operators: ['LIKE'],
+        wildcards: ['startsWith'],
+        transformations: ['UPPER']
+      }
+    },
+    scoringKeys: ['attributes.BldgAbbr', 'attributes.Number', 'attributes.BldgName'],
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
+    popupComponent: Popups.BuildingPopupComponent,
+    searchActive: true
+  },
+  {
+    source: 'building-exact',
+    name: 'Building',
+    url: `${Connections.basemapUrl}/1`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['Number'],
+        operators: ['=']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.BldgName} ({attributes.Number})',
+    popupComponent: Popups.BuildingPopupComponent,
+    searchActive: false
+  },
+  {
+    source: 'university-departments',
+    name: 'University Departments',
+    url: `${Connections.departmentUrl}`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['DeptName', 'CollegeName', 'DeptAbbre'],
+        operators: ['LIKE', 'LIKE', 'LIKE'],
+        wildcards: ['includes', 'includes', 'includes'],
+        transformations: ['UPPER', 'UPPER', 'UPPER']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.DeptName}',
+    popupComponent: Popups.BasePopupComponent,
+    searchActive: true,
+    altLookup: {
+      source: 'building-exact',
+      reference: {
+        keys: ['attributes.HOME1']
+      }
+    }
+  },
+  {
+    source: 'university-departments-exact',
+    name: 'University Departments',
+    url: `${Connections.departmentUrl}`,
+    queryParams: {
+      ...commonQueryParams,
+      resultRecordCount: 100,
+      where: {
+        keys: ['HOME1'],
+        operators: ['=']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.DeptName}',
+    searchActive: false
+  },
+  {
+    source: 'parking-garage',
+    name: 'Parking Garage',
+    url: `${Connections.basemapUrl}/0`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['LotName', 'Name'],
+        operators: ['LIKE', 'LIKE'],
+        wildcards: ['includes', 'includes'],
+        transformations: ['UPPER', 'UPPER']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.LotName}',
+    popupComponent: Popups.BuildingPopupComponent,
+    searchActive: true
+  },
+  {
+    source: 'parking-lot',
+    name: 'Parking Lot',
+    url: `${Connections.basemapUrl}/12`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['LotName'],
+        operators: ['LIKE'],
+        wildcards: ['includes'],
+        transformations: ['UPPER']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.LotName}',
+    popupComponent: Popups.ParkingLotPopupComponent,
+    searchActive: true
+  },
+  {
+    source: 'points-of-interest',
+    name: 'Points of Interest',
+    url: `${Connections.inforUrl}/0`,
+    queryParams: {
+      ...commonQueryParams,
+      where: {
+        keys: ['Name'],
+        operators: ['LIKE'],
+        wildcards: ['includes'],
+        transformations: ['UPPER']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.Name}',
+    popupComponent: Popups.PoiPopupComponent,
+    searchActive: true
+  },
+  {
+    source: 'bike-racks',
+    name: 'Bike Racks',
+    url: `${Connections.bikeRacksUrl}`,
+    queryParams: {
+      ...commonQueryParams,
+      resultRecordCount: 9999,
+      where: {
+        keys: ['1'],
+        operators: ['=']
+      }
+    },
+    featuresLocation: 'features',
+    displayTemplate: '{attributes.Type}',
+    searchActive: false
+  }
+];
+
+export const SelectionSymbols = {
+  polygon: {
+    type: 'simple-fill',
+    style: 'solid',
+    color: [252, 227, 0, 0.55],
+    outline: {
+      color: [252, 227, 0, 0.8],
+      width: '2px'
+    }
+  },
+  point: {
+    type: 'simple-marker',
+    style: 'circle',
+    size: 8,
+    outline: {
+      width: 2
+    }
+  },
+  multipoint: {
+    type: 'simple-marker',
+    style: 'circle',
+    size: 8,
+    outline: {
+      width: 2
+    }
   }
 };
