+++ apps/ues-recycling-data-api-nest/src/environments/environment.dev.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: true,
-  port: 27000,
+  port: 27003,
   globalPrefix: '',
   allowedOrigins: [
     'https://maps.apogee.tamu.edu',
@@ -9,5 +9,5 @@ export const environment = {
   ]
 };
 
-export { productionDbConfig as dbConfig } from './ormconfig';
-export { productionClientConfig as idpConfig } from './oidc-client-config';
+export { devDbConfig as dbConfig } from './ormconfig';
+export { devClientConfig as idpConfig } from './oidc-client-config';
