+++ apps/ues-recycling-data-api-nest/src/environments/environment.prod.ts
@@ -1,8 +1,12 @@
 export const environment = {
   production: true,
-  port: 28052,
+  port: 27003,
   globalPrefix: '',
-  allowedOrigins: ['https://dev.aggiemap.tamu.edu', 'https://ues-dev.geoservices.tamu.edu', 'https://maps.apogee.tamu.edu']
+  allowedOrigins: [
+    'https://maps.apogee.tamu.edu',
+    'https://ues-dev.geoservices.tamu.edu',
+    'https://ues.geoservices.tamu.edu'
+  ]
 };
 
 export { productionDbConfig as dbConfig } from './ormconfig';
